requirements.txt
google-auth==2.16.0
google-api-python-client==2.79.0
python-dotenv==1.0.0
requests==2.31.0
pdfplumber==0.10.3
